Replace err with error, and ensure private macros are private.
The change to error is mainly to reduce confusion with Result::Err.

Additionally, this un-deprecates into_origin, since it serves a different
role than into_error, and can be used without having to manually specify
types.

The as_err macro is renamed to __with_new_errorpoint so as to ensure it
is not advertised as usable to consumers of the crate.